Preserve RSpec configuration across runs
RSpec 2 resets the global configuration and world each time a run completes.
We need to keep the configuration because that’s where a suite’s spec_helper
records its config (don’t want to have to re-evaluate it for each spec).
We need to keep the world because that’s where shared example groups are
kept.

Apparently RSpec 3 works differently in this regard.

Also includes a TODO list with the various behaviors that need automated
tests.