Handle INT and TERM signals in `multirspec`
Closes #3

For INT (^C), we terminate the workers and then print the current progress.

For TERM (the default signal from `kill`) we terminate the workers and do
nothing else.

This commit also changes how workers respond to `TERM`. Previously they
used the default ruby behavior, which converts the signal into a
SignalException. This exception is caught by the RSpec runner, so the
process doesn’t actually die. The workers now immediately die when
receiving TERM. This does mean that the `after` hooks for the current spec
do not run, but by the time we get to sending `TERM`, I don't think we want
them to.